requirements.txt
python-dotenv
ipykernel
jupyter
jupyterlab
pandas
pydub
openai
